workers/requirements.txt
# Article processing worker (DB, LLM, prompts)
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
